requirements.txt
httpx[http2]==0.27.2
cloudscraper==1.2.71
playwright==1.49.0
beautifulsoup4==4.12.3
lxml==5.3.0
feedgen==1.0.0
PyYAML==6.0.2
tenacity==9.0.0
structlog==24.4.0
anyio==4.6.2
pytest==8.3.4
pytest-asyncio==0.24.0

